requirements.txt
langchain-core
langchain-openai
langgraph
langsmith
langchain-community
python-dotenv
openai